pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contractguard"
version = "2.0.0"
description = "Stop bad inputs before they break your systems — 9 analyzers for secrets, PII, SQL, configs, Dockerfiles, dependencies, and more."
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [{name = "ContractGuard Team"}]
keywords = ["security", "secrets", "pii", "sql", "regex", "json", "dockerfile", "dependency", "sarif", "linter", "ci", "sast"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Quality Assurance",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "jsonschema>=4.20.0",
    "sqlparse>=0.5.0",
    "jinja2>=3.1.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
]

[project.scripts]
contractguard = "contractguard.cli:app"

[project.urls]
Homepage = "https://github.com/contractguard/contractguard"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["src/contractguard"]
